requirements.txt

typer[all]==0.12.3
rich==13.7.1
PyYAML==6.0.2
pydantic==2.9.2
